added scenes

+++ assets/gama/vector.h
@@ -3,11 +3,11 @@
 #include <stdlib.h>
 
 typedef struct {
-  float x;
-  float y;
+  double x;
+  double y;
 } Pos;
 
-Pos *at(float x, float y) {
+Pos *at(double x, double y) {
   Pos *p = (Pos *)malloc(sizeof(Pos));
   p->x = x;
   p->y = y;
@@ -31,9 +31,9 @@ Vector *createVector(Pos *pos, Pos *vel, Pos *acc) {
 
 Vector *vectorAt(Pos *pos) { return createVector(pos, at(0, 0), at(0, 0)); }
 
-void updateVector(Vector *v, float theta) {
-  v->pos->x += (v->vel->x * theta) + (0.5f * v->acc->x * theta * theta);
-  v->pos->y += (v->vel->y * theta) + (0.5f * v->acc->y * theta * theta);
+void updateVector(Vector *v, double theta) {
+  v->pos->x += (v->vel->x * theta) + (0.5 * v->acc->x * theta * theta);
+  v->pos->y += (v->vel->y * theta) + (0.5 * v->acc->y * theta * theta);
   v->vel->x += v->acc->x * theta;
   v->vel->y += v->acc->y * theta;
 }
